# BACKEND/pic/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: PIC DIAN
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/xml
paths:
...
  /obtenerPreguntasSeccionIII:
      # binds a127 app logic to a route
      x-swagger-router-controller: preguntas_seccion_iii_i_controller
      post:
        description: Controla la creación de usuarios del sistema
        # used as the method name of the controller
        operationId: obtenerPreguntas_iii_i
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/GenericResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /agregarCuestionario:
      x-swagger-router-controller: cuestionario.controller
      post:
        description: agregar cuestionario
        operationId: Cuestionario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /homologacion:
      x-swagger-router-controller: homologaciones_controller
      post: 
        description: homologaciones
        operationId: homologacion
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /envioPlantillaHomologaciones:
      x-swagger-router-controller: ejemplos.controller
      post:
        description: Envio plantilla para ejemplos
        operationId: envioPlantillaHomologaciones
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /envioPlantillaPic:
      x-swagger-router-controller: ejemplos.controller
      post: 
        description: Envio plantilla para ejemplos
        operationId: envioPlantillaPic
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /envioPlantillaPlanta:
      x-swagger-router-controller: ejemplos.controller
      post: 
        description: Envio plantilla para ejemplos
        operationId: envioPlantillaPlanta
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
  GenericResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string